[relates #4] Create StatisticUtil class

+++ src/main/java/cz/muni/fi/pb138/trafficmap/models/AbstractStatistics.java
@@ -21,6 +21,6 @@ public abstract class AbstractStatistics {
 	private int slightlyInjured;
 
 	@Getter @Setter
-	private int propertyDamage;
+	private double propertyDamage;
 
 }
